Use AccountStore type to manage account data
