+++ frame/evm-balances/src/account_data.rs
@@ -1,0 +1,50 @@
+//! Account balances logic.
+
+use super::*;
+
+/// All balance information for an account.
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Default, RuntimeDebug, MaxEncodedLen, TypeInfo)]
+pub struct AccountData<Balance> {
+	/// Non-reserved part of the balance. There may still be restrictions on this, but it is the
+	/// total pool what may in principle be transferred, reserved and used for tipping.
+	///
+	/// This is the only balance that matters in terms of most operations on tokens. It
+	/// alone is used to determine the balance when in the contract execution environment.
+	pub free: Balance,
+	/// Balance which is reserved and may not be used at all.
+	///
+	/// This can still get slashed, but gets slashed last of all.
+	///
+	/// This balance is a 'reserve' balance that other subsystems use in order to set aside tokens
+	/// that are still 'owned' by the account holder, but which are suspendable.
+	/// This includes named reserve and unnamed reserve.
+	pub reserved: Balance,
+	/// The amount that `free` may not drop below when withdrawing for *anything except transaction
+	/// fee payment*.
+	pub misc_frozen: Balance,
+	/// The amount that `free` may not drop below when withdrawing specifically for transaction
+	/// fee payment.
+	pub fee_frozen: Balance,
+}
+
+impl<Balance: Saturating + Copy + Ord> AccountData<Balance> {
+	/// How much this account's balance can be reduced for the given `reasons`.
+	pub(crate) fn usable(&self, reasons: Reasons) -> Balance {
+		self.free.saturating_sub(self.frozen(reasons))
+	}
+
+    /// The amount that this account's free balance may not be reduced beyond for the given
+    /// `reasons`.
+    pub(crate) fn frozen(&self, reasons: Reasons) -> Balance {
+        match reasons {
+            Reasons::All => self.misc_frozen.max(self.fee_frozen),
+            Reasons::Misc => self.misc_frozen,
+            Reasons::Fee => self.fee_frozen,
+        }
+    }
+
+    /// The total balance in this account including any that is reserved and ignoring any frozen.
+    pub(crate) fn total(&self) -> Balance {
+        self.free.saturating_add(self.reserved)
+    }
+}
